requirements.txt
streamlit
pandas
openpyxl
matplotlib
fpdf2
openai
PyMuPDF